Further source code re-organization
