+++ include/MagnetClasses.hh
@@ -20,7 +20,36 @@
 #include "G4SystemOfUnits.hh"
 
 #include "DetectorConstruction.hh"
-#include "FieldClasses.hh"
+//#include "FieldClasses.hh"
+
+// For the field classes
+#include "G4MagneticField.hh"
+#include "G4Navigator.hh"
+
+/** Magnet field base classes
+ *  (must be on top, as it is used in the MagnetBase)
+ * **/
+
+class FieldBase : public G4MagneticField {
+    // Class that has the navigator and transform
+public:
+    FieldBase(G4ThreeVector centerPoint_in, G4LogicalVolume* fieldLV_in) :
+        centerPoint(centerPoint_in), fieldLV(fieldLV_in) {};
+    virtual void PostInitialize() {
+        SetupTransform();
+    }
+private:
+    G4ThreeVector centerPoint;
+    G4LogicalVolume* fieldLV;
+    static G4Navigator* fNavigator;
+protected:
+    void SetupTransform();
+    G4AffineTransform fGlobalToLocal;
+
+    G4double gradient; // [T/m]
+};
+
+/** Magnet/Object geometry base class **/
 
 class MagnetBase {
 public:
@@ -96,6 +125,9 @@ public:
     void AddSD(); // Adds an SD to the detectorLV
 };
 
+/** Various magnets or objects **/
+
+// PLASMA LENS
 class MagnetPLASMA1 : public MagnetBase {
 public:
     MagnetPLASMA1(G4double zPos_in, G4bool doRelPos_in, G4double length_in, G4double gradient_in,
@@ -110,6 +142,20 @@ private:
     G4double cryHeight;          // [G4 length units]
 };
 
+class FieldPLASMA1 : public FieldBase {
+public:
+    FieldPLASMA1(G4double current_in, G4double radius_in,
+                 G4ThreeVector centerPoint_in, G4LogicalVolume* fieldLV_in);
+    virtual void GetFieldValue(const G4double point[4], G4double field[6]) const;
+
+private:
+    G4double plasmaTotalCurrent; // [A]
+    G4double capRadius;  // [G4 units]
+
+};
+
+// CIRCULAR OPENING COLLIMATOR
+
 class MagnetCOLLIMATOR1 : public MagnetBase {
 public:
     MagnetCOLLIMATOR1(G4double zPos_in, G4bool doRelPos_in, G4double length_in, G4double gradient_in,
@@ -123,6 +169,8 @@ private:
     G4double height = 50.0*mm;  //[G4 length units]
     G4double radius = 50.0*mm;  //[G4 length units]
 };
+
+// TARGET -- JUST A SIMPLE SLAB OF MATERIAL
 
 class MagnetTARGET : public MagnetBase {
 public:
